requirements_face.txt
absl-py==0.9.0
appnope==0.1.0
astor==0.8.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.4.5.2
cycler==0.11.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fonttools==4.38.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.16.1
h5py==2.10.0
importlib-metadata==1.6.1
ipykernel==5.3.0
ipython @ file:///opt/concourse/worker/volumes/live/02237e64-34e3-452e-5c63-14b73548ce0e/volume/ipython_1592348474440/work
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.4.4
MarkupSafe==1.1.1
matplotlib==3.5.3
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mtcnn==0.1.1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.1
opencv-python==4.7.0.72
opt-einsum==3.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.3
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pyzmq==18.1.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
tensorflow==2.0.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typing-extensions==4.5.0
wcwidth==0.2.4
webencodings==0.5.1
wrapt==1.12.1
zipp==3.1.0
